# cache: remove one Clone() from Populate2 modify operation