tCache.Row() clones the model to get 'existing'. The another explicit
Clone() is done to get 'modified' which is passed to ApplyModifications()
and updated with the ovsdb row update. If 'modified' and 'existing'
are different, then 'modified' is passed to Update() which does two more
clones. First for the original row to (presumably?) nothing in the old
row changes underneath Update() while it's processing the update.
Finally Update() clones 'modified' and assigns that as the new row
in the cache.

So, we clone the cache row into 'existing', clone that into 'modified',
update 'modified', clone 'modified' again to update the cache, and then
throw it away. Ideally we could assign 'modified' to the cache instead of
clone it again, but we have to pass it to the event handlers and we have
to ensure it won't be changed underneath the event handler after we've
released the row cache lock.

Since Populate2() owns 'modified' (at least until it hands it off to
the event handler) we can change a few things around and get rid of the
first clone from 'existing' to 'modified'. 'existing'-s only purpose
was to enable the model.Equal() comparison to know if we should actually
update the cache, but since ApplyModifications() makes the changes it
knows when 'modified' is changed. Use that instead of model.Equal().

Lastly, the event handler also used 'existing' but since we're
replacing the original model in the row cache with 'modified' we can
return the original model for the event handler in place of 'existing'
since 'existing' was (previous to this patch) cloned from the original
model anyway.

All this avoids a Clone() in a hotpath. Testcases and benchmarks added.

```
Before (cpu: Intel(R) Core(TM) i7-10610U CPU @ 1.80GHz)
---
BenchmarkPopulate2UpdateArray-8   	      37	  27568442 ns/op
BenchmarkPopulate2UpdateArray-8   	      38	  27635436 ns/op
BenchmarkPopulate2UpdateArray-8   	      40	  27003451 ns/op

BenchmarkPopulate2UpdateArray-8   	      37	  29406288 ns/op
BenchmarkPopulate2UpdateArray-8   	      37	  27819324 ns/op
BenchmarkPopulate2UpdateArray-8   	      43	  27707464 ns/op

BenchmarkPopulate2UpdateArray-8   	      36	  29182002 ns/op
BenchmarkPopulate2UpdateArray-8   	      37	  28766003 ns/op
BenchmarkPopulate2UpdateArray-8   	      39	  27447263 ns/op

BenchmarkPopulate2UpdateArray-8   	      34	  30350489 ns/op
BenchmarkPopulate2UpdateArray-8   	      42	  27639410 ns/op
BenchmarkPopulate2UpdateArray-8   	      39	  27781289 ns/op

After (cpu: Intel(R) Core(TM) i7-10610U CPU @ 1.80GHz)
---
BenchmarkPopulate2UpdateArray-8   	      40	  25367097 ns/op
BenchmarkPopulate2UpdateArray-8   	      49	  25201249 ns/op
BenchmarkPopulate2UpdateArray-8   	      45	  24333118 ns/op

BenchmarkPopulate2UpdateArray-8   	      46	  24455417 ns/op
BenchmarkPopulate2UpdateArray-8   	      46	  25009966 ns/op
BenchmarkPopulate2UpdateArray-8   	      42	  24985884 ns/op

BenchmarkPopulate2UpdateArray-8   	      39	  26141899 ns/op
BenchmarkPopulate2UpdateArray-8   	      44	  25563094 ns/op
BenchmarkPopulate2UpdateArray-8   	      43	  25235681 ns/op

BenchmarkPopulate2UpdateArray-8   	      39	  26473300 ns/op
BenchmarkPopulate2UpdateArray-8   	      44	  25417718 ns/op
BenchmarkPopulate2UpdateArray-8   	      42	  25356652 ns/op
```